Remove zpu_bootram_bd.bmm
Remove the file "zpu_bootram_bd.bmm" because it is generated during
the build and should never have been included in the repository in the
first place - if you have the tools to build the code, you have the
tools that generate it.

Also add a .gitignore file to make sure it stays out of the repo.