Add console mode detection for Wii
Add a detection of Cube vs. Wii mode when GCVideo-DVI is built for the
Wii. This is needed since Nintendo has fixed an off-by-one error in
the hardware so that the Wii uses exactly 48000Hz audio sample rate,
but that fix is disabled in Gamecube mode and thus the sample rate
shifts to 48042Hz there.