Remove the video->audio TIG constraint
According to reports, the version with the TIG constraint suffers from
audio noise on some FPGA boards, so it appears that some path is
under-constrained/not properly synched without it. Removing the TIG
does not cause any timing errors to pop up anymore after splitting the
clock enable counters in the audio subsystem.