Implement a better DCM reset method for Wii
Change the DCM reset procedure so it waits until a few edges are seen
on BClock before releasing the DCM from reset. This delays startup of
the DCM long enough to avoid a period where the Wii main clock
switches to 27MHz after power-on or boot from standby which led to
loss of DCM lock.

TO ensure that this delay is also applied upon power-on, the startup
settings are modified to not wait for DCM lock anymore and the faster
config rate is activated again as the problematic time interval can
now be avoided more accurately.