menu.c: Remove selection of value menuitems
Remove the "value_submenu" code that allowed selection of
value-containing menuitems to change them. Changing the value without
selecting it was always possible (and in fact mandatory for bool-like
values), so it can be removed to save a bit of space and maybe even
reduce user confusion as the interaction method is a bit unusual.