Use table-based scanline profiles instead of a strength setting
Switch the scanline settings from a simple strength value to a
table-based system that uses the pixel luma to determine scanline
strength. The new implementation offers three profiles (plus "off") to
cover multiple use cases and because that happens to be a convenient
amount to store in BRAM or config flash.