Fix VIC for 16:9 video modes
The VIC for 16:9 modes is calculated by incrementing the one for 4:3,
but this was done twice by mistake. This can confuse some sinks
which receive a 480p signal with a VIC indicating a 720p mode, for
example the HDMI analyzer I use chooses its OSD scaling factor using
the VIC if available.